requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
pydantic==1.10.24
python-dotenv==1.1.1
